chore(core): separate function files
